requirements.txt
# Runtime dependencies for the pricing agent and website demo.
openai>=1.99.0
requests>=2.31.0
beautifulsoup4>=4.12.0
APScheduler>=3.10.0
python-dotenv>=1.0.0

# Test dependency.
pytest>=8.0.0
